requirements.txt
yara-python==4.2.3
pyattck==6.0.0
watchdog==2.1.6
python-magic==0.4.27